// app/_types/types.ts
export type NodeType = "readFile" | "summarize" | "email" | "report" | "note";
export type EdgeType = "animated" | "pulse" | "running" | "done" | "error";
// ...
export type NodeStatus = "idle" | "pending" | "running" | "success" | "error";

export type EditorMode = "Editor" | "Executions";

export interface ExecutionLog {
  id: string; //virtual id io not in database just for frontend use
  user_id: string;
  node_id: string;
  workflow_id: string;
  node_label: string;
  status: NodeStatus;
  error_message: string | null;
  output: string | null;
  duration: number | null;
  file_name: string | null;
  finished_at: string;
  started_at: string;
}
// ...
export type UploadedFile = {
  id: string;
  user_id: string;
  workflow_id: string;
  node_id: string | null;
  filename: string;
  original_filename: string;
  file_size: number;
  mime_type: string;
  storage_path: string;
  uploaded_at: string;
};